+++ packages/local-webrtc-browser-sdk/src/http-api/frontend-log.ts
@@ -14,19 +14,14 @@ export const frontendLog = async (options: {
   body: FrontendLog;
 }): Promise<boolean> => {
   const { apiURL, authToken, body } = options;
-  try {
-    const response = await fetch(`${apiURL}/frontend-log`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${authToken}`,
-      },
-      body: JSON.stringify(body),
-    });
+  const response = await fetch(`${apiURL}/frontend-log`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${authToken}`,
+    },
+    body: JSON.stringify(body),
+  });
 
-    return response.status === 201;
-  } catch {
-    // ログの失敗で以降の処理を止めるべきではないため、例外は握りつぶす
-    return false;
-  }
+  return response.status === 201;
 };