+++ packages/local-webrtc-browser-sdk/src/sdk/host-local-webrtc-sdk.ts
@@ -2,7 +2,7 @@ import { ArmdozerId, PilotId } from "gbraver-burst-core";
 import { nanoid } from "nanoid";
 import { Observable } from "rxjs";
 
-import { waitUntilIceCandidate } from "../webrtc/wait-untilIce-candidate";
+import { gatherAllIceCandidates } from "../webrtc/gather-all-ice-candidate";
 import { createRoom } from "../websocket/create-room";
 import { FrontendLogManager } from "./frontend-log-manager";
 import {
@@ -88,15 +88,21 @@ class HostLocalWebRTCSDKImpl implements HostLocalWebRTCSDK {
       const sdp = await connection.createOffer();
 
       await this.#frontendLog.log({ type: "ICE_CANDIDATE_START", spanId });
-      const [iceCandidates] = await Promise.all([
+      const [{ iceCandidates, iceCandidateErrors }] = await Promise.all([
         // icecandidateイベントはsetLocalDescriptionの後に発生するため、先に待機しておく
-        waitUntilIceCandidate(connection),
+        gatherAllIceCandidates(connection),
         connection.setLocalDescription(sdp),
       ]);
       await this.#frontendLog.log({ type: "ICE_CANDIDATE_END", spanId });
 
       const websocket = await this.#websocketConnection.getOrCreate();
-      const roomID = await createRoom({ websocket, sdp, iceCandidates });
+      const [roomID] = await Promise.all([
+        createRoom({ websocket, sdp, iceCandidates }),
+        ...iceCandidateErrors.map((error) =>
+          this.#frontendLog.log({ type: "ICE_CANDIDATE_ERROR", spanId, error }),
+        ),
+      ]);
+
       if (roomID === null) {
         this.#websocketConnection.gracefulDisconnect();
         return null;