+++ packages/ws-signal/src/core/frontend-log/ice-candidate.ts
@@ -1,27 +1,23 @@
 import { z } from "zod";
 
+import { SpanIdContainer, SpanIdContainerSchema } from "./span-id";
+
 /** ログ：ICE候補送信開始 */
-export type ICECandidateStart = {
+export type ICECandidateStart = SpanIdContainer & {
   type: "ICE_CANDIDATE_START";
-  /** スパンID */
-  spanId: string;
 };
 
 /** ICECandidateStart zodスキーマ */
-export const ICECandidateStartSchema = z.object({
+export const ICECandidateStartSchema = SpanIdContainerSchema.extend({
   type: z.literal("ICE_CANDIDATE_START"),
-  spanId: z.string(),
 });
 
 /** ログ：ICE候補送信終了 */
-export type ICECandidateEnd = {
+export type ICECandidateEnd = SpanIdContainer & {
   type: "ICE_CANDIDATE_END";
-  /** スパンID */
-  spanId: string;
 };
 
 /** ICECandidateEnd zodスキーマ */
-export const ICECandidateEndSchema = z.object({
+export const ICECandidateEndSchema = SpanIdContainerSchema.extend({
   type: z.literal("ICE_CANDIDATE_END"),
-  spanId: z.string(),
 });