+++ packages/ws-signal/src/core/frontend-log/index.ts
@@ -7,6 +7,10 @@ import {
   ICECandidateStartSchema,
 } from "./ice-candidate";
 import {
+  ICECandidateError,
+  ICECandidateErrorSchema,
+} from "./ice-candidate-error";
+import {
   SignalingEnd,
   SignalingEndSchema,
   SignalingStart,
@@ -15,12 +19,17 @@ import {
 
 /** フロントエンドログ */
 export type FrontendLog =
-  ICECandidateStart | ICECandidateEnd | SignalingStart | SignalingEnd;
+  | ICECandidateStart
+  | ICECandidateEnd
+  | ICECandidateError
+  | SignalingStart
+  | SignalingEnd;
 
 /** FrontendLog zodスキーマ */
 export const FrontendLogSchema = z.union([
   ICECandidateStartSchema,
   ICECandidateEndSchema,
+  ICECandidateErrorSchema,
   SignalingStartSchema,
   SignalingEndSchema,
 ]);