+++ packages/aws-vpc/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @gbraver-burst-network/aws-vpc
 
+## 1.27.0
+
+### Minor Changes
+
+- フロントエンドログにICE Candidate Errorを追加
+- 5be6fb0: あいことば対戦にフロントエンド用ログを追加
+- 3556f85: ICE収集エラーをフロントエンドログに追加した
+- 70b5b93: ICE Candidate Errorのメッセージを調整
+
+## 1.27.0-beta.2
+
+### Minor Changes
+
+- ICE Candidate Errorのメッセージを調整
+
+## 1.27.0-beta.1
+
+### Minor Changes
+
+- ICE収集エラーをフロントエンドログに追加した
+
+## 1.27.0-beta.0
+
+### Minor Changes
+
+- あいことば対戦にフロントエンド用ログを追加
+
 ## 1.26.3
 
 ### Patch Changes