+++ packages/serverless-stub/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @gbraver-burst-network/serverless-stub
 
+## 1.27.0
+
+### Minor Changes
+
+- フロントエンドログにICE Candidate Errorを追加
+- 5be6fb0: あいことば対戦にフロントエンド用ログを追加
+- 3556f85: ICE収集エラーをフロントエンドログに追加した
+- 70b5b93: ICE Candidate Errorのメッセージを調整
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [5be6fb0]
+- Updated dependencies [3556f85]
+- Updated dependencies [70b5b93]
+  - @gbraver-burst-network/browser-sdk@1.27.0
+
+## 1.27.0-beta.2
+
+### Minor Changes
+
+- ICE Candidate Errorのメッセージを調整
+
+### Patch Changes
+
+- Updated dependencies
+  - @gbraver-burst-network/browser-sdk@1.27.0-beta.2
+
+## 1.27.0-beta.1
+
+### Minor Changes
+
+- ICE収集エラーをフロントエンドログに追加した
+
+### Patch Changes
+
+- Updated dependencies
+  - @gbraver-burst-network/browser-sdk@1.27.0-beta.1
+
+## 1.27.0-beta.0
+
+### Minor Changes
+
+- あいことば対戦にフロントエンド用ログを追加
+
+### Patch Changes
+
+- Updated dependencies
+  - @gbraver-burst-network/browser-sdk@1.27.0-beta.0
+
 ## 1.26.3
 
 ### Patch Changes